chapters int to float -> for inbetween chapters

+++ src/components/PostEdit.tsx
@@ -273,7 +273,7 @@ const PostEdit: FC<PostEditProps> = ({
                                     onChange={(e) =>
                                         setPost({
                                             ...post,
-                                            chapter: parseInt(e.target.value)
+                                            chapter: parseFloat(e.target.value)
                                         })
                                     }
                                 />
